# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion de films en Go.",
        "title": "Movie API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "email": "support@movieapi.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/movies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Renvoie la liste complète des films",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Lister les films",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Movie"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ajoute un nouveau film à la base de données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Créer un film",
                "parameters": [
                    {
                        "description": "Infos du film",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateMovieRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Film créé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erreur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Renvoie les détails d'un film spécifique",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Récupérer un film par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID du film",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Movie"
                        }
                    },
                    "404": {
                        "description": "Film non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations d'un film existant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Modifier un film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID du Film",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles infos du film",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateMovieRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Movie"
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Film non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Efface définitivement un film de la base de données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Supprimer un film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID du Film",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film supprimé avec succès",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Film non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreateMovieRequest": {
            "type": "object",
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Action",
                        "Sci-Fi"
                    ]
                },
                "rating": {
                    "type": "number",
                    "example": 8.7
                },
                "release_year": {
                    "type": "integer",
                    "example": 1999
                },
                "review": {
                    "type": "string",
                    "example": "Un chef d'oeuvre de SF"
                },
                "title": {
                    "type": "string",
                    "example": "The Matrix"
                }
            }
        },
        "main.Movie": {
            "type": "object",
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Action",
                        "Sci-Fi"
                    ]
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "rating": {
                    "type": "number",
                    "example": 8.7
                },
                "release_year": {
                    "type": "integer",
                    "example": 1999
                },
                "review": {
                    "type": "string",
                    "example": "Un chef d'oeuvre de SF"
                },
                "title": {
                    "type": "string",
                    "example": "The Matrix"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}